+++ app/src/main/java/com/example/tomascrd/c_r_s_h/components/DrawableComponent.java
@@ -17,8 +17,8 @@ public abstract class DrawableComponent {
     /**
      * X and Y coordinates
      */
-    protected float xPos;
-    protected float yPos;
+    protected int xPos;
+    protected int yPos;
 
     /**
      * Draws the components on screen
